# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "\"赚圈圈API 返回的goods如果其中的goodsin为yes代表它可以进行交易，即：未下架\"",
        "title": "miniproject",
        "termsOfService": "http://swagger.io/terrms/",
        "contact": {
            "name": "yyj",
            "email": "2105753640@qq.com"
        },
        "version": "1.0.0"
    },
    "host": "119.3.133.235:8080",
    "basePath": "/api/vi",
    "paths": {
        "/money/entrance": {
            "post": {
                "description": "\"登录的api\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login"
                ],
                "summary": "\"用户登录\"",
                "parameters": [
                    {
                        "description": "id 学号 password 密码进行base64加密后的字符串",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.user"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg\":   \"登录成功\",\"token\": token,\"tips\": \"请保留token并将其放在之后的请求头中\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "msg\":\"unauthorization\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "msg\":\"token生成错误\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/money/goods/scanning": {
            "get": {
                "description": "\"点击进入商品详情页的api\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Good"
                ],
                "summary": "\"查询某个商品的详细信息\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商品编号",
                        "name": "goodsid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg\":\"success\",\"infor\":tables.User,\"user\":tables.User,\"tips\":\"如果goodsin是no则代表已经下架，此时则不显示开启交易按钮\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "msg\":\"error happened\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/money/homepage": {
            "get": {
                "description": "\"主页的api\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Good"
                ],
                "summary": "\"主页内容\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg\":\"success\",\"infor\":[]tables.Good}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "msg\":\"error happened in server\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/money/message": {
            "get": {
                "description": "\"消息通知的api\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "\"返回用户购买、收藏后的信息\"",
                "responses": {
                    "200": {
                        "description": "msg\":\"success\",\"infor\":[]tables.Message}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "msg\":\"error happened\",\"infor\":[]tables.Message}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/money/my/cart": {
            "get": {
                "description": "\"购物车\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "My"
                ],
                "summary": "\"查看我的购物车\"",
                "responses": {
                    "200": {
                        "description": "msg\":\"check successfully\",\"infot\":[]tables.Good}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "204": {
                        "description": "msg\":\"check successfully\",\"infot\":\"nothing\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "msg\":\"error happened in server\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/money/my/goods": {
            "get": {
                "description": "\"橱窗\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "My"
                ],
                "summary": "\"查看我的橱窗\"",
                "responses": {
                    "200": {
                        "description": "msg\":\"check successfully\",\"infot\":[]tables.Good}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "msg\":\"error happened in server\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/money/my/goods/finish": {
            "get": {
                "description": "\"点击确认完成时的api\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trade"
                ],
                "summary": "\"商家完成订单\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商品编号",
                        "name": "goodsid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg\":\"success\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "msg\":\"error happened in server\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/money/my/goods/unfish": {
            "get": {
                "description": "\"返回订单,需要点完成的是‘my sell’-\u003e[]string 是与我做交易的人的id,因为一个商品可能被多个人购买，所以string切片的长度就是‘完成订单’的订单数,点评价的是‘my buy’\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trade"
                ],
                "summary": "\"返回用户与卖家未完成的订单\"",
                "responses": {
                    "200": {
                        "description": "msg\":\"success\",\"my buy\":[]tables.Good,\"my sell\":[]Two}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error happened\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/money/my/message": {
            "get": {
                "description": "\"我的个人信息的api\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "My"
                ],
                "summary": "\"返回我的信息\"",
                "responses": {
                    "200": {
                        "description": "msg\":\"avatar 是头像对应的url\",\"infor\":tables.User}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "msg\":\"error happened in server\",\"infor\":tables.User}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.user": {
            "type": "object",
            "required": [
                "id",
                "password"
            ],
            "properties": {
                "id": {
                    "description": "一定要输入的加上了required",
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    }
}